Create lets upgrade assignment 1
In this assignment, we have added a <header> tag to define the header section of the webpage. Inside the <header> tag, there is an <h1> heading element that represents the header text.
The CSS styles defined in the <style> block within the <head> section set the background color of the header to light blue (#87CEEB) using the background-color property. The text-align: center; property centers the text within the header. The padding: 20px; property adds some padding around the header content.
The <main> tag is used to define the main content section of the webpage. We can add our actual content within the <main> tag.